requirements.txt
black
imageio
imageio-ffmpeg
ipdb
jupyter
lpips
matplotlib
numpy
opencv-python
pillow
plotly
scipy
tqdm
trimesh
